Rename the name of header file to avoid error
