+++ carta/cpp/core/Data/Image/PCacheSqlite3Vector.cpp
@@ -1,4 +1,4 @@
-﻿#include "PCacheSqlite3.h"
+﻿#include "PCacheSqlite3Vector.h"
 #include "CartaLib/Hooks/GetPersistentCache.h"
