+++ carta/cpp/core/core.pro
@@ -84,7 +84,6 @@ HEADERS += \
     Data/Image/Save/SaveService.h \
     Data/Image/Save/SaveView.h \
     Data/Image/Save/SaveViewLayered.h \
-    Data/Image/PCacheSqlite3.h \
     Data/Image/hashidsxx/hashids.h \
     Data/Selection.h \
     Data/Layout/Layout.h \
@@ -175,6 +174,7 @@ HEADERS += \
     Hacks/ManagedLayerView.h \
     Hacks/LayeredViewDemo.h \
     Hacks/InteractiveShapes.h \
+    Data/Image/PCacheSqlite3Vector.h
 
 SOURCES += \
     Viewer.cpp \
@@ -233,7 +233,6 @@ SOURCES += \
     Data/Image/Save/SaveService.cpp \
     Data/Image/Save/SaveView.cpp \
     Data/Image/Save/SaveViewLayered.cpp \
-    Data/Image/PCacheSqlite3.cpp \
     Data/Image/hashidsxx/hashids.cpp \
     Data/DataLoader.cpp \
     Data/Error/ErrorReport.cpp \
@@ -333,6 +332,7 @@ SOURCES += \
     Hacks/ManagedLayerView.cpp \
     Hacks/LayeredViewDemo.cpp \
     Hacks/InteractiveShapes.cpp \
+    Data/Image/PCacheSqlite3Vector.cpp
 
 #message( "common            PWD=$$PWD")
 #message( "common         IN_PWD=$$IN_PWD")
